backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
requests
httpx
pytest